removed first create project configuration
